added XRays, drafted comments for fix for issue 1
